兼容旧的nsq采集方法
