+++ agent/metrics_agent.go
@@ -55,7 +55,6 @@ import (
 	_ "flashcat.cloud/categraf/inputs/nginx"
 	_ "flashcat.cloud/categraf/inputs/nginx_upstream_check"
 	_ "flashcat.cloud/categraf/inputs/nsq"
-	_ "flashcat.cloud/categraf/inputs/nsq_slim"
 	_ "flashcat.cloud/categraf/inputs/ntp"
 	_ "flashcat.cloud/categraf/inputs/nvidia_smi"
 	_ "flashcat.cloud/categraf/inputs/oracle"
