--- conf/input.nsq_slim/nsq_slim.toml
@@ -1,20 +1,0 @@
-# # collect interval
-# interval = 15
-
-[[instances]]
-## The Nsq API URI used to collect statistical information.
-# url = "http://127.0.0.1:8171/api/topics"
- 
-# headers={Authorization="", X-Forwarded-For="", Host=""}
- 
-# timeout="5s"
-
-# # basic auth
-# username=""
-# password=""
-
-## append some labels for series
-# labels = { region="shanghai", product="nsq" }
-
-## interval = global.interval * interval_times
-# interval_times = 1
